==> notebooks/01_Scrapeo_continuado.ipynb <==
from bs4 import BeautifulSoup
import random
# Requests
import requests

import pandas as pd
import numpy as np

from time import sleep

# Importar librerías para automatización de navegadores web con Selenium
from selenium.webdriver.common.by import By
from selenium import webdriver  # Selenium es una herramienta para automatizar la interacción con navegadores web.
from webdriver_manager.chrome import ChromeDriverManager  # ChromeDriverManager gestiona la instalación del controlador de Chrome.
from selenium.webdriver.common.keys import Keys  # Keys es útil para simular eventos de teclado en Selenium.
from selenium.webdriver.support.ui import Select  # Select se utiliza para interactuar con elementos <select> en páginas web.
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import NoSuchElementException # Excepciones comunes de selenium que nos podemos encontrar 

driver = webdriver.Chrome()
url_proyecto = "https://minimalismbrand.com/collections/sudaderas"
driver.get(url_proyecto)
# colocar prints
driver.maximize_window()
driver.implicitly_wait(5)

sudaderas = driver.find_element('xpath', '/html/body/main/div/div[5]/div/div[2]/div[3]/div[2]/div/div[1]/div/div/div[1]/a/div/div[1]/div/div/div/img').click()

# Realizar un pequeño scroll hacia abajo
driver.execute_script("window.scrollBy(0, 600);")

wait = WebDriverWait(driver, 3)
titulo = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'title')))

# Obtener y mostrar el nombre del producto
nombre_producto = titulo.text
print(nombre_producto)

precio = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'current-price')))
precio_producto = precio.text
print(precio_producto)

# Esperar a que el color esté presente y obtener su valor
color = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'label')))
color_producto = color.text
print(color_producto)

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_Desc = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div')))
descripciones = ac_Desc.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div/p[2]').text

# Mostrar la descripción del producto
print(f"Descripción del producto: {descripciones}")

driver.execute_script("window.scrollBy(0, 250);")

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_desglose = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/div/div')))
costes = ac_desglose.find_elements("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/div/div/ul/li')

lista_desglose = []
lista_columnas = []
for i in costes:
    texto = i.text
    porcentaje = texto.split(' — ')[0]
    columna = texto.split(' — ')[1]
    lista_desglose.append(porcentaje)
    lista_columnas.append(columna)

lista_desglose

lista_columnas

df_sudadera = pd.DataFrame({
    'nombre':nombre_producto,
    'precio': precio_producto,
    'descripción': descripciones,
    lista_columnas[0]:[lista_desglose[0]],
    lista_columnas[1]:[lista_desglose[1]],
    lista_columnas[2]:[lista_desglose[2]],
    lista_columnas[3]:[lista_desglose[3]],
})
df_sudadera

driver.close()

Sudadera capucha

driver = webdriver.Chrome()
url_proyecto = "https://minimalismbrand.com/collections/sudaderas"
driver.get(url_proyecto)
# colocar prints
driver.maximize_window()
driver.implicitly_wait(5)

driver.execute_script("window.scrollBy(0, 600);")


sudaderas_capucha = driver.find_element('xpath', '/html/body/main/div/div[5]/div/div[2]/div[3]/div[2]/div/div[8]/div/div/div[1]/a/div/div[1]/div/div/div/img').click()

# Realizar un pequeño scroll hacia abajo
driver.execute_script("window.scrollBy(0, 600);")

wait = WebDriverWait(driver, 3)
titulo = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'title')))

# Obtener y mostrar el nombre del producto
nombre_producto = titulo.text
print(nombre_producto)

precio = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'current-price')))
precio_producto = precio.text
print(precio_producto)

# Esperar a que el color esté presente y obtener su valor
color = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'label')))
color_producto = color.text
print(color_producto)

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_Desc = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div')))
descripciones = ac_Desc.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div/p[2]').text

# Mostrar la descripción del producto
print(f"Descripción del producto: {descripciones}")

driver.execute_script("window.scrollBy(0, 250);")

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_desglose = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/div/div')))
costes = ac_desglose.find_elements("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[12]/div/details/div/div/ul/li')

lista_desglose = []
lista_columnas = []
for i in costes:
    texto = i.text
    porcentaje = texto.split(' — ')[0]
    columna = texto.split(' — ')[1]
    lista_desglose.append(porcentaje)
    lista_columnas.append(columna)

lista_desglose

lista_columnas

df_sudadera_cap = pd.DataFrame({
    'nombre':nombre_producto,
    'precio': precio_producto,
    'descripción': descripciones,
    lista_columnas[0]:[lista_desglose[0]],
    lista_columnas[1]:[lista_desglose[1]],
    lista_columnas[2]:[lista_desglose[2]],
    lista_columnas[3]:[lista_desglose[3]],
})
df_sudadera_cap

# Jerseys

driver = webdriver.Chrome()
url_proyecto = "https://minimalismbrand.com/collections/sweater-men"
driver.get(url_proyecto)
# colocar prints
driver.maximize_window()
driver.implicitly_wait(5)

driver.execute_script("window.scrollBy(0, 600);")


sudaderas_jersey = driver.find_element('xpath', '/html/body/main/div/div[5]/div/div[2]/div[3]/div[2]/div/div[1]/div/div/div[1]/a/div/div[1]/div/div/div/img').click()

# Realizar un pequeño scroll hacia abajo
driver.execute_script("window.scrollBy(0, 600);")

wait = WebDriverWait(driver, 3)
titulo = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'title')))

# Obtener y mostrar el nombre del producto
nombre_producto = titulo.text
print(nombre_producto)

precio = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'current-price')))
precio_producto = precio.text
print(precio_producto)

# Esperar a que el color esté presente y obtener su valor
color = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'label')))
color_producto = color.text
print(color_producto)

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_Desc = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div')))
descripciones = ac_Desc.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div/p[1]').text

# Mostrar la descripción del producto
print(f"Descripción del producto: {descripciones}")

driver.execute_script("window.scrollBy(0, 900);")

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_desglose = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div')))
costes = ac_desglose.find_elements("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div/ul/li')

lista_desglose = []
lista_columnas = []
for i in costes:
    texto = i.text
    porcentaje = texto.split(' — ')[0]
    columna = texto.split(' — ')[1]
    lista_desglose.append(porcentaje)
    lista_columnas.append(columna)
lista_desglose
lista_columnas
df_jersey_punto_lana = pd.DataFrame({
    'nombre':nombre_producto,
    'precio': precio_producto,
    'descripción': descripciones,
    lista_columnas[0]:[lista_desglose[0]],
    lista_columnas[1]:[lista_desglose[1]],
    lista_columnas[2]:[lista_desglose[2]],
    lista_columnas[3]:[lista_desglose[3]],
})
df_jersey_punto_lana

driver.close()

## Jersey punto fino

driver = webdriver.Chrome()
url_proyecto = "https://minimalismbrand.com/collections/sweater-men"
driver.get(url_proyecto)
# colocar prints
driver.maximize_window()
driver.implicitly_wait(5)

driver.execute_script("window.scrollBy(0, 1200);")


jersey_punto_fino = driver.find_element('xpath', '/html/body/main/div/div[5]/div/div[2]/div[3]/div[2]/div/div[5]/div/div/div[1]/a/div/div[1]/div/div/div/img').click()

# Realizar un pequeño scroll hacia abajo
driver.execute_script("window.scrollBy(0, 600);")

wait = WebDriverWait(driver, 3)
titulo = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'title')))

# Obtener y mostrar el nombre del producto
nombre_producto = titulo.text
print(nombre_producto)

precio = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'current-price')))
precio_producto = precio.text
print(precio_producto)

# Esperar a que el color esté presente y obtener su valor
color = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'label')))
color_producto = color.text
print(color_producto)

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_Desc = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div')))
descripciones = ac_Desc.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div/p[1]').text

# Mostrar la descripción del producto
print(f"Descripción del producto: {descripciones}")

driver.execute_script("window.scrollBy(0, 900);")

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_desglose = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div')))
costes = ac_desglose.find_elements("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div/ul/li')

lista_desglose = []
lista_columnas = []
for i in costes:
    texto = i.text
    porcentaje = texto.split(' — ')[0]
    columna = texto.split(' — ')[1]
    lista_desglose.append(porcentaje)
    lista_columnas.append(columna)
lista_desglose
lista_columnas
df_jersey_punto_fino = pd.DataFrame({
    'nombre':nombre_producto,
    'precio': precio_producto,
    'descripción': descripciones,
    lista_columnas[0]:[lista_desglose[0]],
    lista_columnas[1]:[lista_desglose[1]],
    lista_columnas[2]:[lista_desglose[2]],
    lista_columnas[3]:[lista_desglose[3]],
})
df_jersey_punto_fino


## Jersey lana merino


driver = webdriver.Chrome()
url_proyecto = "https://minimalismbrand.com/collections/sweater-men"
driver.get(url_proyecto)
# colocar prints
driver.maximize_window()
driver.implicitly_wait(5)

driver.execute_script("window.scrollBy(0, 1800);")


jersey_lana_merino = driver.find_element('xpath', '/html/body/main/div/div[5]/div/div[2]/div[3]/div[2]/div/div[10]/div/div/div[1]/a/div/div[1]/div/div/div/img').click()

# Realizar un pequeño scroll hacia abajo
driver.execute_script("window.scrollBy(0, 600);")

wait = WebDriverWait(driver, 3)
titulo = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'title')))

# Obtener y mostrar el nombre del producto
nombre_producto = titulo.text
print(nombre_producto)

precio = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'current-price')))
precio_producto = precio.text
print(precio_producto)

# Esperar a que el color esté presente y obtener su valor
color = wait.until(EC.presence_of_element_located((By.CLASS_NAME, 'label')))
color_producto = color.text
print(color_producto)

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_Desc = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div')))
descripciones = ac_Desc.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[9]/div/details/div/div/p[1]').text

# Mostrar la descripción del producto
print(f"Descripción del producto: {descripciones}")

driver.execute_script("window.scrollBy(0, 300);")

sleep(2)
try: 
    driver.find_element("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/summary').click()
except:
    print('No encuentro el botón')

ac_desglose = wait.until(EC.presence_of_element_located(("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div')))
costes = ac_desglose.find_elements("xpath", '/html/body/main/div/div[1]/div/div[2]/div[2]/div/div[11]/div/details/div/div/ul/li')

lista_desglose = []
lista_columnas = []
for i in costes:
    texto = i.text
    porcentaje = texto.split(' — ')[0]
    columna = texto.split(' — ')[1]
    lista_desglose.append(porcentaje)
    lista_columnas.append(columna)
lista_desglose
lista_columnas
df_jersey_lana_merino = pd.DataFrame({
    'nombre':nombre_producto,
    'precio': precio_producto,
    'descripción': descripciones,
    lista_columnas[0]:[lista_desglose[0]],
    lista_columnas[1]:[lista_desglose[1]],
    lista_columnas[2]:[lista_desglose[2]],
    lista_columnas[3]:[lista_desglose[3]],
})
df_jersey_lana_merino

df_completo = pd.concat([df_sudadera, df_sudadera_cap, df_jersey_punto_lana, df_jersey_punto_fino, df_jersey_lana_merino], ignore_index=True)

df_completo
