==> notebooks/03_all_productos.ipynb <==
# Web Entera minimalism


import pandas as pd
ruta_base = "../datos/02_Products/"

# Cargar archivos CSV en variables
accesorios = pd.read_csv(ruta_base + "Accesorios.csv")
banador = pd.read_csv(ruta_base + "Bañador.csv")
calcetines = pd.read_csv(ruta_base + "Calcetines.csv")
calzoncillos_hombre = pd.read_csv(ruta_base + "Calzoncillos_hombre.csv")
camisetas_minimalism_hombre = pd.read_csv(ruta_base + "Camisetas_minimalism_hombre.csv")
carteras = pd.read_csv(ruta_base + "Carteras.csv")
gorros = pd.read_csv(ruta_base + "Gorros.csv")
guantes = pd.read_csv(ruta_base + "Guantes.csv")
jerseis = pd.read_csv(ruta_base + "Jerseis.csv")
pantalones_hombre = pd.read_csv(ruta_base + "pantalones_hombre.csv")
pantalones_mujer = pd.read_csv(ruta_base + "Pantalones_mujer.csv")
ropa_interior_mujer = pd.read_csv(ruta_base + "RopaInteriorMujer.csv")
sudaderas = pd.read_csv(ruta_base + "Sudaderas.csv")


accesorios = accesorios.rename(columns={"product-block__title":"producto","product-price__item":"precio (€)"})
accesorios["precio (€)"] = accesorios["precio (€)"].str.strip("€").astype(float)
accesorios["categoria"] = "accesorios"
accesorios = accesorios[["categoria","producto","precio (€)"]]
accesorios["talla"] = ["única","única",['13/14"','15/16"'],"única"]
accesorios


banador["precio (€)"] = 55.00
banador["categoria"] = "bañadores"
banador = banador[["categoria","product-block__title","precio (€)"]]
banador = banador.rename(columns={"product-block__title":"producto"})
banador["talla"] = [["XS","S","M","L","XL","XXL"]]*len(banador)
banador

calcetines = calcetines.rename(columns={"product-block__title":"producto","product-price__item":"precio (€)"})
calcetines["categoria"] = "calcetines"
calcetines["precio (€)"] = calcetines["precio (€)"].str.strip("€").astype(float)
calcetines = calcetines[["categoria","producto","precio (€)"]]
calcetines["talla"] = [["35-40","41-46"]]*len(calcetines)
calcetines

calzoncillos_hombre.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)
calzoncillos_hombre["categoria"] = "calzoncillos"
calzoncillos_hombre = calzoncillos_hombre[["categoria", "producto", "precio (€)"]]
calzoncillos_hombre["precio (€)"] = calzoncillos_hombre["precio (€)"].str.strip("€").astype(float)
calzoncillos_hombre["talla"] = [["S","M","L","XL","XXL"]]*len(calzoncillos_hombre)
calzoncillos_hombre

camisetas_minimalism_hombre.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
camisetas_minimalism_hombre["categoria"] = "camisetas y polos"

# Reordenar las columnas
camisetas_minimalism_hombre = camisetas_minimalism_hombre[["categoria", "producto", "precio (€)"]]
camisetas_minimalism_hombre["producto"] = camisetas_minimalism_hombre["producto"].str.replace("\n", "", regex=True).str.strip()
# Limpiar la columna de precio y convertir a float
camisetas_minimalism_hombre["precio (€)"] = camisetas_minimalism_hombre["precio (€)"].str.strip("€").astype(float)
camisetas_minimalism_hombre["talla"] = [["XS","S","M","L","XL","XXL"]]*len(camisetas_minimalism_hombre)
camisetas_minimalism_hombre.sample()

carteras.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
carteras["categoria"] = "carteras"

# Reordenar las columnas
carteras = carteras[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
carteras["precio (€)"] = carteras["precio (€)"].str.strip("€").astype(float)
carteras["producto"] = carteras["producto"].str.replace("\n", "", regex=True).str.strip()

carteras["talla"] = "única"
carteras

gorros.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
gorros["categoria"] = "gorros"

# Reordenar las columnas
gorros = gorros[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
gorros["precio (€)"] = gorros["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
gorros["producto"] = gorros["producto"].str.replace("\n", "", regex=True).str.strip()
gorros["talla"] = "única"
gorros

guantes.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
guantes["categoria"] = "guantes"

# Reordenar las columnas
guantes = guantes[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
guantes["precio (€)"] = guantes["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
guantes["producto"] = guantes["producto"].str.replace("\n", "", regex=True).str.strip()
guantes["talla"] = "única"
guantes

jerseis.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
jerseis["categoria"] = "jerseis"

# Reordenar las columnas
jerseis = jerseis[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
jerseis["precio (€)"] = jerseis["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
jerseis["producto"] = jerseis["producto"].str.replace("\n", "", regex=True).str.strip()
tallas_1 = ["S", "M", "L", "XL", "XXL"]  # Para los primeros 8 productos
tallas_2 = ["XS", "S", "M", "L", "XL", "XXL"]  # Para los siguientes 2 productos

# Asignar la columna 'talla' en base a la longitud del DataFrame
jerseis["talla"] = [tallas_1] * 8 + [tallas_2] * 2
jerseis


pantalones_hombre.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
pantalones_hombre["categoria"] = "pantalones_hombre"

# Reordenar las columnas
pantalones_hombre = pantalones_hombre[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
pantalones_hombre["precio (€)"] = pantalones_hombre["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
pantalones_hombre["producto"] = pantalones_hombre["producto"].str.replace("\n", "", regex=True).str.strip()

tallas_1 = ["38","40","42","44","46"]  # Para los primeros 8 productos
tallas_2 = ["XS", "S", "M", "L", "XL"]  # Para los siguientes 2 productos

# Asignar la columna 'talla' en base a la longitud del DataFrame
pantalones_hombre["talla"] = [tallas_1] * 2 + [tallas_2] * 7
pantalones_hombre

pantalones_mujer.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
pantalones_mujer["categoria"] = "pantalones_mujer"

# Reordenar las columnas
pantalones_mujer = pantalones_mujer[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
pantalones_mujer["precio (€)"] = pantalones_mujer["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
pantalones_mujer["producto"] = pantalones_mujer["producto"].str.replace("\n", "", regex=True).str.strip()

tallas_1 = ["36","38","40","42"]  # Para los primeros 8 productos
tallas_2 = ["XS", "S", "M", "L", "XL"]  # Para los siguientes 2 productos

# Asignar la columna 'talla' en base a la longitud del DataFrame
pantalones_mujer["talla"] = [tallas_1] * 2 + [tallas_2] * 7
pantalones_mujer

ropa_interior_mujer.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
ropa_interior_mujer["categoria"] = "ropa_interior_mujer"

# Reordenar las columnas
ropa_interior_mujer = ropa_interior_mujer[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
ropa_interior_mujer["precio (€)"] = ropa_interior_mujer["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
ropa_interior_mujer["producto"] = ropa_interior_mujer["producto"].str.replace("\n", "", regex=True).str.strip()
ropa_interior_mujer = ropa_interior_mujer.loc[0:12]

tallas_1 = ["38","40","42","44","46"]  # Para los primeros 8 productos
tallas_2 = ["XS", "S", "M", "L", "XL","XXL"]  # Para los siguientes 2 productos

# Asignar la columna 'talla' en base a la longitud del DataFrame
ropa_interior_mujer["talla"] = [["XS", "S", "M", "L", "XL","XXL"]] * len(ropa_interior_mujer)
ropa_interior_mujer

sudaderas.rename(columns={
    "product-block__title": "producto",
    "product-price__item": "precio (€)"
}, inplace=True)

# Añadir la columna 'Categoria'
sudaderas["categoria"] = "sudaderas"

# Reordenar las columnas
sudaderas = sudaderas[["categoria", "producto", "precio (€)"]]

# Limpiar la columna de precio y convertir a float
sudaderas["precio (€)"] = sudaderas["precio (€)"].str.strip("€").astype(float)

# Eliminar saltos de línea y espacios adicionales en la columna 'producto'
sudaderas["producto"] = sudaderas["producto"].str.replace("\n", "", regex=True).str.strip()
sudaderas["talla"] = [["XS","S","M","L","XL","XXL"]]*len(sudaderas)
sudaderas

df = pd.concat([accesorios,banador,calcetines,calzoncillos_hombre,camisetas_minimalism_hombre,carteras,gorros,guantes,jerseis,
                pantalones_hombre,pantalones_mujer,ropa_interior_mujer,sudaderas],ignore_index=True)
df

df.to_csv("../datos/09_fullproducts_minimalism.csv")
